+++ src/bg_clip.h
@@ -48,8 +48,8 @@ struct TankCollectionData {
 };
 
 void BGClipMotherBrainUpdateGlass(u8 bg, u16 value, u16 yPosition, u16 xPosition);
-void BGClipSetBG1BlockValue(u16 value, u16 yPosition, u16 xPosition);
-void BGClipSetRawBG1BlockValue(u16 value, u16 yPosition, u16 xPosition);
+void BGClipSetBG1BlockValue(u32 value, u16 yPosition, u16 xPosition);
+void BGClipSetRawBG1BlockValue(u32 value, u16 yPosition, u16 xPosition);
 void BGClipSetClipdataBlockValue(u16 value, u16 yPosition, u16 xPosition);
 void BGClipCheckTouchingSpecialClipdata(void);
 void BGClipApplyClipdataChangingTransparency(void);
